requirements.txt
pip
streamlit
streamlit-modal
requests
jinja2
langchain-groq
langchain-core
langchain
langchain-community
langchain_text_splitters
faiss-cpu
transformers
sentence-transformers
python-dotenv
torch
reportlab
pydantic
graphviz
langgraph
PyPDF2
PyMuPDF
selenium
webdriver-manager
beautifulsoup4
lxml